requirements.txt
scrapy==2.10.1
scrapy-redis==0.7.3
pip==23.2.1
pika==1.3.2
